Move buildAdminOwnerProps to a dedicated trait (Concerns\BuildsAdminOwnerProps)
Address Copilot review: the base Controller is a general-purpose class
inherited by every controller, so adding a Trip-specific method there
violates SRP. Move the method into a dedicated trait and use it only in
the two controllers that need it (MapController, TripController).
