+++ app/Http/Controllers/Controller.php
@@ -2,30 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Trip;
-
 abstract class Controller
 {
-    /**
-     * Build owner props for the admin banner when an admin views another user's trip.
-     *
-     * @return array{owner?: array{id: int, name: string}}
-     */
-    protected function buildAdminOwnerProps(Trip $trip): array
-    {
-        $user = auth()->user();
-
-        if (! $user->isAdmin() || $trip->user_id === $user->id) {
-            return [];
-        }
-
-        $trip->loadMissing('user:id,name');
-
-        return [
-            'owner' => [
-                'id' => $trip->user->id,
-                'name' => $trip->user->name,
-            ],
-        ];
-    }
+    //
 }
