+++ app/Http/Controllers/MapController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\Concerns\BuildsAdminOwnerProps;
 use App\Models\Trip;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Inertia\Inertia;
@@ -10,6 +11,7 @@ use Inertia\Response;
 class MapController extends Controller
 {
     use AuthorizesRequests;
+    use BuildsAdminOwnerProps;
 
     public function show(Trip $trip): Response
     {
